refactor(frontend): don't let vite inject css
