+++ frontend/src/components/countdown/countdown.tsx
@@ -6,8 +6,8 @@ import {
   useStore,
   useStylesScoped$,
 } from "@builder.io/qwik";
-import styles from "./countdown.css";
 import { intervalToDuration } from "date-fns";
+import styles from "./countdown.css?inline";
 
 export const START_DATE = new Date("2022-10-01T00:00:00+07:00").getTime(); // 1st of October 2022 UTC+7
 export const END_DATE = new Date("2022-10-31T00:00:00+07:00").getTime(); // 31st of October 2022 UTC+7
